Switch to use Solr
We want to keep existing URLs and since we don't have a feature flag we now have to rename relevant routes, controllers and views.

Renames old routes, controllers and views by adding `os_` prefix and removes `solr_` prefix.

Models and Services are kept intact.

Also adds missing translations and _sort partial.
