+++ app/views/os_datasets/_meta_data.html.erb
@@ -1,7 +1,8 @@
 <% content_for :head do %>
-  <%= tag.meta name: 'dc:title',   content: @dataset.title %>
+  <%= tag.meta name: 'dc:title',   content: unescape(@dataset.title) %>
   <%= tag.meta name: 'dc:creator', content: @dataset.organisation.title %>
-  <%= tag.meta name: 'dc:date', content: Date.parse(@dataset.public_updated_at).to_formatted_s %>
+  <%= tag.meta name: 'dc:date',    content: Date.parse(@dataset.public_updated_at).to_formatted_s %>
+
   <% if @dataset.licence_title.present? %>
     <%= tag.meta name: 'dc:rights', content: @dataset.licence_title %>
   <% end %>
@@ -20,13 +21,14 @@
     <div class="column-two-thirds">
       <div class="dgu-metadata__box dgu-metadata__box--in-dataset">
         <dl class="metadata">
-          <% unless @dataset.datafiles.present? %>
+          <% unless @dataset.released %>
             <dt><%= t('.availability') %>:</dt>
             <dd>
               <span class="dgu-highlight">Not released</span>
             </dd>
           <% end %>
           <dt><%= t('.published_by') %>:</dt>
+
           <dd property="dc:creator">
             <%= @dataset.organisation.title %>
           </dd>
@@ -38,23 +40,18 @@
 
           <dt><%= t('.topic') %>:</dt>
           <% if @dataset.topic %>
-            <dd><%= @dataset.topic %></dd>
+            <dd><%= @dataset.topic['title'] %></dd>
           <% else %>
             <dd><%= t('.topic_not_added') %></dd>
           <% end %>
 
           <dt><%= t('.licence') %>:</dt>
           <dd property="dc:rights">
-            <% if Search::Solr::OGL_IDS.any? { |id| id === @dataset.licence_code } %>
-              <%= link_to t('.uk_ogl'),
-                          t('.uk_ogl_url'),
-                            rel: 'dc:rights',
-                            class: 'govuk-link' %>
-            <% elsif @dataset.licence_url.present? %>
+            <% if @dataset.licence_url.present? %>
               <%= link_to @dataset.licence_title,
-                          @dataset.licence_url.html_safe,
-                            rel: 'dc:rights',
-                            class: 'govuk-link' %>
+                          @dataset.licence_url,
+                          rel: 'dc:rights',
+                          class: 'govuk-link' %>
             <% elsif @dataset.licence_title.present? %>
               <%= @dataset.licence_title %>
             <% elsif @dataset.licence_custom.present? %>
@@ -80,16 +77,30 @@
         </div>
       </div>
     </div>
+
     <div class="column-one-third dgu-dataset-right">
       <div class="dgu-dataset-right__sidebar__publisher_datasets">
         <h3 class="heading-small"><%= t('.publisher_datasets') %></h3>
-        <%= link_to search_solr_path(filters: { publisher: @dataset.organisation.title }), class: 'govuk-link' do %>
+        <%= link_to search_os_path(filters: { publisher: @dataset.organisation.title }), class: 'govuk-link' do %>
           <%= "All datasets from #{@dataset.organisation.title}" %>
         <% end %>
       </div>
 
+      <% unless @related_datasets.empty? %>
+        <div class="dgu-dataset-right__sidebar">
+          <h3 class="heading-small"><%= t('.related_datasets') %></h3>
+          <ul>
+            <% @related_datasets.each do |dataset| %>
+              <li>
+                <%= link_to unescape(dataset.title), os_dataset_path(dataset.uuid, dataset.name), class: 'govuk-link' %>
+              </li>
+            <% end %>
+          </ul>
+        </div>
+      <% end %>
+
       <h3 class="heading-small"><%= t('.search_gov_data') %></h3>
-      <form action="/search/solr" method="GET" class="dgu-search-box">
+      <form action="/search" method="GET" class="dgu-search-box">
         <label for="q" class="visuallyhidden"><%= t('.accessibility.search_box_label') %></label>
         <input id="q"
                name="q"
@@ -97,7 +108,7 @@
                class="form-control dgu-search-box__input"/><button type="submit"
                                                                    class="dgu-search-box__button">
                                                             <%= t('.accessibility.search_box_button')%>
-                                                          </button>
+                                                           </button>
       </form>
     </div>
   </div>
