+++ app/views/os_search/_sort.html.erb
@@ -1,0 +1,24 @@
+<div class="dgu-sort">
+  <%= render "govuk_publishing_components/components/select", {
+    id: "sort-datasets",
+    name: "sort",
+    label: t('.sort_by'),
+    options: [
+      {
+        text: t('.match'),
+        value: 'best',
+        selected: @sort === 'best' || @sort != 'recent'
+      },
+      {
+        text: t('.recent'),
+        value: 'recent',
+        selected: @sort === 'recent'
+      }
+    ]
+  } %>
+  <div class="govuk-visually-hidden">
+    <%= render "govuk_publishing_components/components/button", {
+      text: t('.accessibility.submit_button'),
+    } %>
+  </div>
+</div>
