+++ app/views/search/_no_results.html.erb
