--- app/views/solr_datasets/_no_datafiles.html.erb
@@ -1,12 +1,0 @@
-<div class="govuk-details__text govuk-!-margin-bottom-4">
-  <p class="govuk-body"><%= t('solr_datasets.show.not_released') %></p>
-  <p class="govuk-body">
-    <% if contact_information_exists?(dataset) %>
-      <%= t('solr_datasets.show.contact_the_publisher') %>
-    <% else %>
-      <%= t('solr_datasets.show.contact_the_team') %>
-      <%= link_to 'data.gov.uk/support', support_path, class: 'govuk-link' %>
-      <%= t('solr_datasets.show.if_you_have_questions') %>
-    <% end %>
-  </p>
-</div>
