+++ config/routes.rb
@@ -48,11 +48,11 @@ Rails.application.routes.draw do
   get "dataset/:uuid/:name", to: "datasets#show", as: "dataset"
   get "dataset/:dataset_uuid/:name/datafile/:datafile_uuid/preview", to: "previews#show", as: "datafile_preview"
 
-  # /solr feature flag
-  get "search/solr", to: "solr_search#search"
-  get "dataset/solr/:uuid", to: "solr_datasets#show", uuid: /[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}/
-  get "dataset/solr/:uuid/:name", to: "solr_datasets#show", as: "solr_dataset"
-  get "dataset/solr/:dataset_uuid/:name/datafile/:datafile_uuid/preview", to: "solr_previews#show", as: "solr_datafile_preview"
+  # Old OpenSearch routes
+  get "search/os", to: "os_search#search"
+  get "dataset/os/:uuid", to: "os_datasets#show", uuid: /[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}/
+  get "dataset/os/:uuid/:name", to: "os_datasets#show", as: "os_dataset"
+  get "dataset/os/:dataset_uuid/:name/datafile/:datafile_uuid/preview", to: "os_previews#show", as: "os_datafile_preview"
 
   get "acknowledge", to: "messages#acknowledge"
